+++ src/classes/redis-connection.ts
@@ -12,6 +12,7 @@ import {
   isRedisInstance,
   isRedisVersionLowerThan,
 } from '../utils';
+import { version } from '../version';
 import * as scripts from '../scripts';
 
 const overrideMessage = [
@@ -195,13 +196,17 @@ export class RedisConnection extends EventEmitter {
     return this.initializing;
   }
 
-  protected loadCommands(providedScripts?: Record<string, RawCommand>): void {
+  protected loadCommands(
+    version: string,
+    providedScripts?: Record<string, RawCommand>,
+  ): void {
     const finalScripts =
       providedScripts || (scripts as Record<string, RawCommand>);
     for (const property in finalScripts as Record<string, RawCommand>) {
       // Only define the command if not already defined
-      if (!(<any>this._client)[finalScripts[property].name]) {
-        (<any>this._client).defineCommand(finalScripts[property].name, {
+      const commandName = `${finalScripts[property].name}:${version}`;
+      if (!(<any>this._client)[commandName]) {
+        (<any>this._client).defineCommand(commandName, {
           numberOfKeys: finalScripts[property].keys,
           lua: finalScripts[property].content,
         });
@@ -225,7 +230,7 @@ export class RedisConnection extends EventEmitter {
 
     await RedisConnection.waitUntilReady(this._client);
 
-    this.loadCommands();
+    this.loadCommands(version);
 
     if (this._client['status'] !== 'end') {
       this.version = await this.getRedisVersion();