+++ src/classes/scripts.ts
@@ -36,14 +36,18 @@ import {
 import { ErrorCode } from '../enums';
 import { array2obj, getParentKey, isRedisVersionLowerThan } from '../utils';
 import { ChainableCommander } from 'ioredis';
-
+import { version } from '../version';
 export type JobData = [JobJsonRaw | number, string?];
 
 export class Scripts {
+  protected version;
+
   moveToFinishedKeys: (string | undefined)[];
 
   constructor(protected queue: MinimalQueue) {
     const queueKeys = this.queue.keys;
+
+    this.version = version;
 
     this.moveToFinishedKeys = [
       queueKeys.wait,
@@ -63,13 +67,22 @@ export class Scripts {
     ];
   }
 
+  public execCommand(
+    client: RedisClient | ChainableCommander,
+    commandName: string,
+    args: any[],
+  ) {
+    const commandNameWithVersion = `${commandName}:${this.version}`;
+    return (<any>client)[commandNameWithVersion](args);
+  }
+
   async isJobInList(listKey: string, jobId: string): Promise<boolean> {
     const client = await this.queue.client;
     let result;
     if (isRedisVersionLowerThan(this.queue.redisVersion, '6.0.6')) {
-      result = await (<any>client).isJobInList([listKey, jobId]);
+      result = await this.execCommand(client, 'isJobInList', [listKey, jobId]);
     } else {
-      result = await (<any>client).lpos(listKey, jobId);
+      result = await client.lpos(listKey, jobId);
     }
     return Number.isInteger(result);
   }
@@ -91,8 +104,7 @@ export class Scripts {
     ];
 
     keys.push(pack(args), job.data, encodedOpts);
-
-    return (<any>client).addDelayedJob(keys);
+    return this.execCommand(client, 'addDelayedJob', keys);
   }
 
   protected addPrioritizedJob(
@@ -114,8 +126,7 @@ export class Scripts {
     ];
 
     keys.push(pack(args), job.data, encodedOpts);
-
-    return (<any>client).addPrioritizedJob(keys);
+    return this.execCommand(client, 'addPrioritizedJob', keys);
   }
 
   protected addParentJob(
@@ -133,8 +144,7 @@ export class Scripts {
     ];
 
     keys.push(pack(args), job.data, encodedOpts);
-
-    return (<any>client).addParentJob(keys);
+    return this.execCommand(client, 'addParentJob', keys);
   }
 
   protected addStandardJob(
@@ -156,8 +166,7 @@ export class Scripts {
     ];
 
     keys.push(pack(args), job.data, encodedOpts);
-
-    return (<any>client).addStandardJob(keys);
+    return this.execCommand(client, 'addStandardJob', keys);
   }
 
   async addJob(
@@ -211,7 +220,7 @@ export class Scripts {
 
     if (parentOpts.waitChildrenKey) {
       result = await this.addParentJob(client, job, encodedOpts, args);
-    } else if (opts.delay) {
+    } else if (typeof opts.delay == 'number') {
       result = await this.addDelayedJob(client, job, encodedOpts, args);
     } else if (opts.priority) {
       result = await this.addPrioritizedJob(client, job, encodedOpts, args);
@@ -257,8 +266,7 @@ export class Scripts {
     const client = await this.queue.client;
 
     const args = this.pauseArgs(pause);
-
-    return (<any>client).pause(args);
+    return this.execCommand(client, 'pause', args);
   }
 
   protected addRepeatableJobArgs(
@@ -298,8 +306,7 @@ export class Scripts {
       opts,
       legacyCustomKey,
     );
-
-    return (<any>client).addRepeatableJob(args);
+    return this.execCommand(client, 'addRepeatableJob', args);
   }
 
   async addJobScheduler(
@@ -315,8 +322,7 @@ export class Scripts {
       queueKeys.delayed,
     ];
     const args = [nextMillis, pack(opts), jobSchedulerId, queueKeys['']];
-
-    return (<any>client).addJobScheduler(keys.concat(args));
+    return this.execCommand(client, 'addJobScheduler', keys.concat(args));
   }
 
   async updateRepeatableJobMillis(
@@ -331,7 +337,7 @@ export class Scripts {
       customKey,
       legacyCustomKey,
     ];
-    return (<any>client).updateRepeatableJobMillis(args);
+    return this.execCommand(client, 'updateRepeatableJobMillis', args);
   }
 
   async updateJobSchedulerNextMillis(
@@ -382,8 +388,7 @@ export class Scripts {
       repeatConcatOptions,
       repeatJobKey,
     );
-
-    return (<any>client).removeRepeatable(args);
+    return this.execCommand(client, 'removeRepeatable', args);
   }
 
   async removeJobScheduler(jobSchedulerId: string): Promise<number> {
@@ -395,24 +400,33 @@ export class Scripts {
 
     const args = [jobSchedulerId, queueKeys['']];
 
-    return (<any>client).removeJobScheduler(keys.concat(args));
+    return this.execCommand(client, 'removeJobScheduler', keys.concat(args));
+  }
+
+  protected removeArgs(
+    jobId: string,
+    removeChildren: boolean,
+  ): (string | number)[] {
+    const keys: (string | number)[] = ['', 'meta'].map(name =>
+      this.queue.toKey(name),
+    );
+
+    const args = [jobId, removeChildren ? 1 : 0];
+
+    return keys.concat(args);
   }
 
   async remove(jobId: string, removeChildren: boolean): Promise<number> {
     const client = await this.queue.client;
 
-    const keys: (string | number)[] = ['', 'meta'].map(name =>
-      this.queue.toKey(name),
-    );
-    const result = await (<any>client).removeJob(
-      keys.concat([jobId, removeChildren ? 1 : 0]),
-    );
+    const args = this.removeArgs(jobId, removeChildren);
+    const result = await this.execCommand(client, 'removeJob', args);
 
-    if (result == ErrorCode.JobBelongsToJobScheduler) {
+    if (result < 0) {
       throw this.finishedErrors({
-        code: ErrorCode.JobBelongsToJobScheduler,
+        code: result,
         jobId,
-        command: 'remove',
+        command: 'removeJob',
       });
     }
 
@@ -433,7 +447,7 @@ export class Scripts {
       duration,
       jobId,
     ];
-    return (<any>client).extendLock(args);
+    return this.execCommand(client, 'extendLock', args);
   }
 
   async updateData<T = any, R = any, N extends string = string>(
@@ -445,7 +459,11 @@ export class Scripts {
     const keys = [this.queue.toKey(job.id)];
     const dataJson = JSON.stringify(data);
 
-    const result = await (<any>client).updateData(keys.concat([dataJson]));
+    const result = await this.execCommand(
+      client,
+      'updateData',
+      keys.concat([dataJson]),
+    );
 
     if (result < 0) {
       throw this.finishedErrors({
@@ -469,7 +487,9 @@ export class Scripts {
     ];
     const progressJson = JSON.stringify(progress);
 
-    const result = await (<any>client).updateProgress(
+    const result = await this.execCommand(
+      client,
+      'updateProgress',
       keys.concat([jobId, progressJson]),
     );
 
@@ -494,7 +514,9 @@ export class Scripts {
       this.queue.toKey(jobId) + ':logs',
     ];
 
-    const result = await (<any>client).addLog(
+    const result = await this.execCommand(
+      client,
+      'addLog',
       keys.concat([jobId, logRow, keepLogs ? keepLogs : '']),
     );
 
@@ -581,7 +603,7 @@ export class Scripts {
   ) {
     const client = await this.queue.client;
 
-    const result = await (<any>client).moveToFinished(args);
+    const result = await this.execCommand(client, 'moveToFinished', args);
     if (result < 0) {
       throw this.finishedErrors({
         code: result,
@@ -593,49 +615,6 @@ export class Scripts {
       if (typeof result !== 'undefined') {
         return raw2NextJobData(result);
       }
-    }
-  }
-
-  finishedErrors({
-    code,
-    jobId,
-    parentKey,
-    command,
-    state,
-  }: {
-    code: number;
-    jobId?: string;
-    parentKey?: string;
-    command: string;
-    state?: string;
-  }): Error {
-    switch (code) {
-      case ErrorCode.JobNotExist:
-        return new Error(`Missing key for job ${jobId}. ${command}`);
-      case ErrorCode.JobLockNotExist:
-        return new Error(`Missing lock for job ${jobId}. ${command}`);
-      case ErrorCode.JobNotInState:
-        return new Error(
-          `Job ${jobId} is not in the ${state} state. ${command}`,
-        );
-      case ErrorCode.JobPendingDependencies:
-        return new Error(`Job ${jobId} has pending dependencies. ${command}`);
-      case ErrorCode.ParentJobNotExist:
-        return new Error(`Missing key for parent job ${parentKey}. ${command}`);
-      case ErrorCode.JobLockMismatch:
-        return new Error(
-          `Lock mismatch for job ${jobId}. Cmd ${command} from ${state}`,
-        );
-      case ErrorCode.ParentJobCannotBeReplaced:
-        return new Error(
-          `The parent job ${parentKey} cannot be replaced. ${command}`,
-        );
-      case ErrorCode.JobBelongsToJobScheduler:
-        return new Error(
-          `Job ${jobId} belongs to a job scheduler and cannot be removed directly. ${command}`,
-        );
-      default:
-        return new Error(`Unknown code ${code} error for ${jobId}. ${command}`);
     }
   }
 
@@ -659,7 +638,7 @@ export class Scripts {
     const client = await this.queue.client;
     const args = this.drainArgs(delayed);
 
-    return (<any>client).drain(args);
+    return this.execCommand(client, 'drain', args);
   }
 
   private removeChildDependencyArgs(
@@ -682,7 +661,11 @@ export class Scripts {
     const client = await this.queue.client;
     const args = this.removeChildDependencyArgs(jobId, parentKey);
 
-    const result = await (<any>client).removeChildDependency(args);
+    const result = await this.execCommand(
+      client,
+      'removeChildDependency',
+      args,
+    );
 
     switch (result) {
       case 0:
@@ -726,7 +709,7 @@ export class Scripts {
     const client = await this.queue.client;
     const args = this.getRangesArgs(types, start, end, asc);
 
-    return (<any>client).getRanges(args);
+    return await this.execCommand(client, 'getRanges', args);
   }
 
   private getCountsArgs(types: JobType[]): (string | number)[] {
@@ -746,7 +729,7 @@ export class Scripts {
     const client = await this.queue.client;
     const args = this.getCountsArgs(types);
 
-    return (<any>client).getCounts(args);
+    return await this.execCommand(client, 'getCounts', args);
   }
 
   protected getCountsPerPriorityArgs(
@@ -768,7 +751,7 @@ export class Scripts {
     const client = await this.queue.client;
     const args = this.getCountsPerPriorityArgs(priorities);
 
-    return (<any>client).getCountsPerPriority(args);
+    return await this.execCommand(client, 'getCountsPerPriority', args);
   }
 
   moveToCompletedArgs<T = any, R = any, N extends string = string>(
@@ -821,7 +804,9 @@ export class Scripts {
       return this.queue.toKey(key);
     });
 
-    return (<any>client).isFinished(
+    return this.execCommand(
+      client,
+      'isFinished',
       keys.concat([jobId, returnValue ? '1' : '']),
     );
   }
@@ -843,16 +828,16 @@ export class Scripts {
     });
 
     if (isRedisVersionLowerThan(this.queue.redisVersion, '6.0.6')) {
-      return (<any>client).getState(keys.concat([jobId]));
+      return this.execCommand(client, 'getState', keys.concat([jobId]));
     }
-    return (<any>client).getStateV2(keys.concat([jobId]));
+    return this.execCommand(client, 'getStateV2', keys.concat([jobId]));
   }
 
   async changeDelay(jobId: string, delay: number): Promise<void> {
     const client = await this.queue.client;
 
     const args = this.changeDelayArgs(jobId, delay);
-    const result = await (<any>client).changeDelay(args);
+    const result = await this.execCommand(client, 'changeDelay', args);
     if (result < 0) {
       throw this.finishedErrors({
         code: result,
@@ -889,7 +874,8 @@ export class Scripts {
     const client = await this.queue.client;
 
     const args = this.changePriorityArgs(jobId, priority, lifo);
-    const result = await (<any>client).changePriority(args);
+
+    const result = await this.execCommand(client, 'changePriority', args);
     if (result < 0) {
       throw this.finishedErrors({
         code: result,
@@ -994,7 +980,7 @@ export class Scripts {
     const client = await this.queue.client;
 
     const args = this.isMaxedArgs();
-    return !!(await (<any>client).isMaxed(args));
+    return !!(await this.execCommand(client, 'isMaxed', args));
   }
 
   async moveToDelayed(
@@ -1007,7 +993,8 @@ export class Scripts {
     const client = await this.queue.client;
 
     const args = this.moveToDelayedArgs(jobId, timestamp, token, delay, opts);
-    const result = await (<any>client).moveToDelayed(args);
+
+    const result = await this.execCommand(client, 'moveToDelayed', args);
     if (result < 0) {
       throw this.finishedErrors({
         code: result,
@@ -1037,7 +1024,11 @@ export class Scripts {
     const client = await this.queue.client;
 
     const args = this.moveToWaitingChildrenArgs(jobId, token, opts);
-    const result = await (<any>client).moveToWaitingChildren(args);
+    const result = await this.execCommand(
+      client,
+      'moveToWaitingChildren',
+      args,
+    );
 
     switch (result) {
       case 0:
@@ -1064,7 +1055,7 @@ export class Scripts {
     const client = await this.queue.client;
 
     const args = this.getRateLimitTtlArgs(maxJobs);
-    return (<any>client).getRateLimitTtl(args);
+    return this.execCommand(client, 'getRateLimitTtl', args);
   }
 
   /**
@@ -1079,7 +1070,7 @@ export class Scripts {
   ): Promise<string[]> {
     const client = await this.queue.client;
 
-    return (<any>client).cleanJobsInSet([
+    return this.execCommand(client, 'cleanJobsInSet', [
       this.queue.toKey(set),
       this.queue.toKey('events'),
       this.queue.toKey('repeat'),
@@ -1150,7 +1141,7 @@ export class Scripts {
 
     const args = this.moveJobsToWaitArgs(state, count, timestamp);
 
-    return (<any>client).moveJobsToWait(args);
+    return this.execCommand(client, 'moveJobsToWait', args);
   }
 
   async promoteJobs(count = 1000): Promise<number> {
@@ -1158,7 +1149,7 @@ export class Scripts {
 
     const args = this.moveJobsToWaitArgs('delayed', count, Number.MAX_VALUE);
 
-    return (<any>client).moveJobsToWait(args);
+    return this.execCommand(client, 'moveJobsToWait', args);
   }
 
   /**
@@ -1198,7 +1189,11 @@ export class Scripts {
       state,
     ];
 
-    const result = await (<any>client).reprocessJob(keys.concat(args));
+    const result = await this.execCommand(
+      client,
+      'reprocessJob',
+      keys.concat(args),
+    );
 
     switch (result) {
       case 1:
@@ -1242,7 +1237,9 @@ export class Scripts {
       }),
     ];
 
-    const result = await (<any>client).moveToActive(
+    const result = await this.execCommand(
+      client,
+      'moveToActive',
       (<(string | number | boolean | Buffer)[]>keys).concat(args),
     );
 
@@ -1266,7 +1263,7 @@ export class Scripts {
 
     const args = [this.queue.toKey(''), jobId];
 
-    const code = await (<any>client).promote(keys.concat(args));
+    const code = await this.execCommand(client, 'promote', keys.concat(args));
     if (code < 0) {
       throw this.finishedErrors({
         code,
@@ -1314,7 +1311,7 @@ export class Scripts {
 
     const args = this.moveStalledJobsToWaitArgs();
 
-    return (<any>client).moveStalledJobsToWait(args);
+    return this.execCommand(client, 'moveStalledJobsToWait', args);
   }
 
   /**
@@ -1345,7 +1342,11 @@ export class Scripts {
 
     const args = [jobId, token, this.queue.toKey(jobId)];
 
-    const pttl = await (<any>client).moveJobFromActiveToWait(keys.concat(args));
+    const pttl = await this.execCommand(
+      client,
+      'moveJobFromActiveToWait',
+      keys.concat(args),
+    );
 
     return pttl < 0 ? 0 : pttl;
   }
@@ -1359,7 +1360,11 @@ export class Scripts {
     ];
     const args = [opts.count, opts.force ? 'force' : null];
 
-    const result = await (<any>client).obliterate(keys.concat(args));
+    const result = await this.execCommand(
+      client,
+      'obliterate',
+      keys.concat(args),
+    );
     if (result < 0) {
       switch (result) {
         case -1:
@@ -1413,9 +1418,12 @@ export class Scripts {
         args.push(1);
       }
 
-      [cursor, offset, items, total, rawJobs] = await (<any>client).paginate(
+      [cursor, offset, items, total, rawJobs] = await this.execCommand(
+        client,
+        'paginate',
         keys.concat(args),
       );
+
       page = page.concat(items);
 
       if (rawJobs && rawJobs.length) {
@@ -1450,6 +1458,49 @@ export class Scripts {
         total,
         jobs,
       };
+    }
+  }
+
+  finishedErrors({
+    code,
+    jobId,
+    parentKey,
+    command,
+    state,
+  }: {
+    code: number;
+    jobId?: string;
+    parentKey?: string;
+    command: string;
+    state?: string;
+  }): Error {
+    switch (code) {
+      case ErrorCode.JobNotExist:
+        return new Error(`Missing key for job ${jobId}. ${command}`);
+      case ErrorCode.JobLockNotExist:
+        return new Error(`Missing lock for job ${jobId}. ${command}`);
+      case ErrorCode.JobNotInState:
+        return new Error(
+          `Job ${jobId} is not in the ${state} state. ${command}`,
+        );
+      case ErrorCode.JobPendingDependencies:
+        return new Error(`Job ${jobId} has pending dependencies. ${command}`);
+      case ErrorCode.ParentJobNotExist:
+        return new Error(`Missing key for parent job ${parentKey}. ${command}`);
+      case ErrorCode.JobLockMismatch:
+        return new Error(
+          `Lock mismatch for job ${jobId}. Cmd ${command} from ${state}`,
+        );
+      case ErrorCode.ParentJobCannotBeReplaced:
+        return new Error(
+          `The parent job ${parentKey} cannot be replaced. ${command}`,
+        );
+      case ErrorCode.JobBelongsToJobScheduler:
+        return new Error(
+          `Job ${jobId} belongs to a job scheduler and cannot be removed directly. ${command}`,
+        );
+      default:
+        return new Error(`Unknown code ${code} error for ${jobId}. ${command}`);
     }
   }
 }