+++ src/utils.ts
@@ -10,6 +10,9 @@ import { ChildMessage, RedisClient } from './interfaces';
 import { EventEmitter } from 'events';
 import * as semver from 'semver';
 
+import { join } from 'path';
+import { readFileSync } from 'fs';
+
 export const errorObject: { [index: string]: any } = { value: null };
 
 export function tryCatch(
@@ -212,6 +215,20 @@ export const parseObjectValues = (obj: {
   return accumulator;
 };
 
+const getCircularReplacer = (rootReference: any) => {
+  const references = new WeakSet();
+  references.add(rootReference);
+  return (_: string, value: any) => {
+    if (typeof value === 'object' && value !== null) {
+      if (references.has(value)) {
+        return '[Circular]';
+      }
+      references.add(value);
+    }
+    return value;
+  };
+};
+
 export const errorToJSON = (value: any): Record<string, any> => {
   const error: Record<string, any> = {};
 
@@ -219,7 +236,7 @@ export const errorToJSON = (value: any): Record<string, any> => {
     error[propName] = value[propName];
   });
 
-  return error;
+  return JSON.parse(JSON.stringify(error, getCircularReplacer(value)));
 };
 
 const INFINITY = 1 / 0;